Big changes to IP: 1.  Renamed James IP with timing InvasionPercolationTimed, and gave the IP basis algorithm the name InvasionPercolation.  The basis/standard algorithm should get the base name. 2.  The basic IP algorithm is not VASTLY faster.  a 200,000 pore network takes just 20 seconds on my laptop
